make make.sh and README.md more informative
make.sh:
  - omit pointless -p option
  - check for improper invocation instead of
    just passing down arguments to lower level tools
  - emit useful error messages
README.md:
  - provide comprehensive guidance to determine which
    header to build with

+++ make.sh
@@ -4,9 +4,18 @@ cd "`dirname $0`"
 export GOPATH="$PWD/go"
 protogendir="$GOPATH/src/parsefuse/protogen"
 
+if ! [ $# -eq 1 ]; then
+    echo "Usage: $0 <path-to-fuse-header>"
+    exit 1
+fi
+if ! [ -f "$1" ]; then
+    echo "$1 is not a regular file"
+    exit 1
+fi
+
 mkdir -p "$protogendir"
 
-ruby -Iruby ruby/gauguin.rb -m messages.yaml $* < protogen.erb | gofmt > "$protogendir/protogen.go"
+ruby -Iruby ruby/gauguin.rb -m messages.yaml -p "$1" < protogen.erb | gofmt > "$protogendir/protogen.go"
 
 go get github.com/ugorji/go/codec
 go install parsefuse
